TrustEVM Staking Contract v1
All the required implementations are included in this commit. The contract will become upgradable in a later commit, towards deployment.
One of the features that was changed from the original SushiSwap design is calculating the 'owed' rewards when the user deposits. These 'owed' rewards are substracted to RewardDebt in the Withdraw function, in order to send them over to the user.