Improve contract builds, contract testing, and CICD.
Now build two versions of the contract (one with test actions and one
without) in the same build directory automatically. Also build the
unit_test for the contracts as well.

That means a path within the Leap build directory must be provided to
the cmake command. Previously that was with eosio_DIR but now the README
and CICD scripts have been updated to use leap_DIR instead.

The change also removes the need for the WITH_TEST_ACTIONS and
WITH_LARGE_STACK CMake options, which have now been removed.
Note that WITH_LOGTIME still exists as an option but it only impacts
the version of the contract that is built with test actions.

A new BUILD_TESTS CMake option which defaults to ON. If set to OFF,
the contract tests will not be built which means there is no requirement
for the leap_DIR option to be provided to cmake. It also means that the
version of the contract with test actions will not be built.

Functionality has been added to shard the consensus tests that
execute via arguments passed along to the test via unit_test.
This allows ctest to execute the consensus tests in parallel, but
parameters that determine how to shard the tests must be set in the
CMakeList.txt file.
Additionally, a convenient consensus_tests label has been assigned
to all shards of the consensus tests so that it is easy to just run the
consensus tests or to just run all tests other than the consensus tests.

The CICD process has been simplified to not require two variants of the
Contract Build job.