Implement gas fee feature.
setingressfee action replaced by more general setfeeparams action
which also allows changing the gas_price and the newly introduced
miner_cut within the config singleton.

The init action now also takes an additional argument which is the same
argument that the setfeeparams action takes. This enables the gas_price
and miner_cut to be set during initialization (and in fact values for
both must be provided). The ingress_bridge_fee does not need to be
provided during init; if it is not provided it will go with the default
value of zero ingress bridge fee.

The pushtx action has been modified to take miner as its first argument
instead of ram_payer. The type is the same but its semantic meaning has
now changed. The RAM costs are now always paid by the contract. And
therefore no authorizer is required on the pushtx action. The purpose
of the miner argument is to specify the EOS account that will receive
the miner portion of the gas fees paid for the EVM transaction.
The miner portion is determined by miner_cut which acts as a percentage
between 0% and 100%.

Additionally, the pushtx action now enforces that the transaction
provided as an argument to the action sets a gas price that is no less
than the gas price set in the contract state.

Unit tests have been added for the gas fee feature.

Some minor tweaks were made to the bridging tests so that their checks
properly account for a non-zero gas price provided in the test
transactions.

Additionally an extra test case has been added to the unit tests for
the block mapping and timestamp changes.

Changes have not yet been made to the integration tests to reflect the
change to the signature of the contract actions.