Use common code with EVM node and contract to fill in block header fields.
Set beneficiary in block header to the special reserved address.
This requires the EVM node to know the name of the account the EVM
contract is deployed to. This is now communicated using the genesis
header nonce.

The previous meaning of the genesis header nonce was the block time
interval. That is now hardcoded to be 1 second.