Working through dependency issues.
Need to link TestHarness into typical python dist-packages. This is taken care of automatically in the install in leap main but not in v4.0.4 being used here.

TestHarness has a dependency on cleos which currently is not installed in the leap-dev package, so need leap package for now as well.